ARROW-12703: [JS] Separate Table from DataFrame
