+++ js/test/unit/table-tests.ts
@@ -17,18 +17,15 @@
 
 import '../jest-extensions';
 import {
-    predicate,
     Data, Schema, Field, Table, RecordBatch, Column,
     Vector, Int32Vector, Float32Vector, Utf8Vector, DictionaryVector,
-    Struct, Float32, Int32, Dictionary, Utf8, Int8, Type
+    Struct, Float32, Int32, Dictionary, Utf8, Int8
 } from '../Arrow';
 import { arange } from './utils';
 
-const { col, lit, custom, and, or, And, Or } = predicate;
-
 const NAMES = ['f32', 'i32', 'dictionary'] as (keyof TestDataSchema)[];
 const F32 = 0, I32 = 1, DICT = 2;
-const test_data = [
+export const test_data = [
     {
         name: `single record batch`,
         table: getSingleRecordBatchTable,
@@ -319,42 +316,6 @@ describe(`Table`, () => {
                 expect(clone).toEqualTable(table);
             });
 
-            describe(`scan()`, () => {
-                test(`yields all values`, () => {
-                    const table = datum.table();
-                    let expected_idx = 0;
-                    table.scan((idx, batch) => {
-                        const columns = batch.schema.fields.map((_, i) => batch.getChildAt(i)!);
-                        expect(columns.map((c) => c.get(idx))).toEqual(values[expected_idx++]);
-                    });
-                });
-                test(`calls bind function with every batch`, () => {
-                    const table = datum.table();
-                    let bind = jest.fn();
-                    table.scan(() => { }, bind);
-                    for (let batch of table.chunks) {
-                        expect(bind).toHaveBeenCalledWith(batch);
-                    }
-                });
-            });
-            describe(`scanReverse()`, () => {
-                test(`yields all values`, () => {
-                    const table = datum.table();
-                    let expected_idx = values.length;
-                    table.scanReverse((idx, batch) => {
-                        const columns = batch.schema.fields.map((_, i) => batch.getChildAt(i)!);
-                        expect(columns.map((c) => c.get(idx))).toEqual(values[--expected_idx]);
-                    });
-                });
-                test(`calls bind function with every batch`, () => {
-                    const table = datum.table();
-                    let bind = jest.fn();
-                    table.scanReverse(() => { }, bind);
-                    for (let batch of table.chunks) {
-                        expect(bind).toHaveBeenCalledWith(batch);
-                    }
-                });
-            });
             test(`count() returns the correct length`, () => {
                 const table = datum.table();
                 const values = datum.values();
@@ -366,151 +327,10 @@ describe(`Table`, () => {
                 expect(table.getColumnIndex('f32')).toEqual(F32);
                 expect(table.getColumnIndex('dictionary')).toEqual(DICT);
             });
+
             const table = datum.table();
             const values = datum.values();
-            let get_i32: (idx: number) => number, get_f32: (idx: number) => number;
-            const filter_tests = [
-                {
-                    name: `filter on f32 >= 0`,
-                    filtered: table.filter(col('f32').ge(0)),
-                    expected: values.filter((row) => row[F32] >= 0)
-                }, {
-                    name: `filter on 0 <= f32`,
-                    filtered: table.filter(lit(0).le(col('f32'))),
-                    expected: values.filter((row) => 0 <= row[F32])
-                }, {
-                    name: `filter on i32 <= 0`,
-                    filtered: table.filter(col('i32').le(0)),
-                    expected: values.filter((row) => row[I32] <= 0)
-                }, {
-                    name: `filter on 0 >= i32`,
-                    filtered: table.filter(lit(0).ge(col('i32'))),
-                    expected: values.filter((row) => 0 >= row[I32])
-                }, {
-                    name: `filter on f32 < 0`,
-                    filtered: table.filter(col('f32').lt(0)),
-                    expected: values.filter((row) => row[F32] < 0)
-                }, {
-                    name: `filter on i32 > 1 (empty)`,
-                    filtered: table.filter(col('i32').gt(0)),
-                    expected: values.filter((row) => row[I32] > 0)
-                }, {
-                    name: `filter on f32 <= -.25 || f3 >= .25`,
-                    filtered: table.filter(col('f32').le(-.25).or(col('f32').ge(.25))),
-                    expected: values.filter((row) => row[F32] <= -.25 || row[F32] >= .25)
-                }, {
-                    name: `filter on !(f32 <= -.25 || f3 >= .25) (not)`,
-                    filtered: table.filter(col('f32').le(-.25).or(col('f32').ge(.25)).not()),
-                    expected: values.filter((row) => !(row[F32] <= -.25 || row[F32] >= .25))
-                }, {
-                    name: `filter method combines predicates (f32 >= 0 && i32 <= 0)`,
-                    filtered: table.filter(col('i32').le(0)).filter(col('f32').ge(0)),
-                    expected: values.filter((row) => row[I32] <= 0 && row[F32] >= 0)
-                }, {
-                    name: `filter on dictionary == 'a'`,
-                    filtered: table.filter(col('dictionary').eq('a')),
-                    expected: values.filter((row) => row[DICT] === 'a')
-                }, {
-                    name: `filter on 'a' == dictionary (commutativity)`,
-                    filtered: table.filter(lit('a').eq(col('dictionary'))),
-                    expected: values.filter((row) => row[DICT] === 'a')
-                }, {
-                    name: `filter on dictionary != 'b'`,
-                    filtered: table.filter(col('dictionary').ne('b')),
-                    expected: values.filter((row) => row[DICT] !== 'b')
-                }, {
-                    name: `filter on f32 >= i32`,
-                    filtered: table.filter(col('f32').ge(col('i32'))),
-                    expected: values.filter((row) => row[F32] >= row[I32])
-                }, {
-                    name: `filter on f32 <= i32`,
-                    filtered: table.filter(col('f32').le(col('i32'))),
-                    expected: values.filter((row) => row[F32] <= row[I32])
-                }, {
-                    name: `filter on f32*i32 > 0 (custom predicate)`,
-                    filtered: table.filter(custom(
-                        (idx: number) => (get_f32(idx) * get_i32(idx) > 0),
-                        (batch: RecordBatch) => {
-                            get_f32 = col('f32').bind(batch);
-                            get_i32 = col('i32').bind(batch);
-                        })),
-                    expected: values.filter((row) => (row[F32] as number) * (row[I32] as number) > 0)
-                }, {
-                    name: `filter out all records`,
-                    filtered: table.filter(lit(1).eq(0)),
-                    expected: []
-                }
-            ];
-            for (let this_test of filter_tests) {
-                const { name, filtered, expected } = this_test;
-                describe(name, () => {
-                    test(`count() returns the correct length`, () => {
-                        expect(filtered.count()).toEqual(expected.length);
-                    });
-                    describe(`scan()`, () => {
-                        test(`iterates over expected values`, () => {
-                            let expected_idx = 0;
-                            filtered.scan((idx, batch) => {
-                                const columns = batch.schema.fields.map((_, i) => batch.getChildAt(i)!);
-                                expect(columns.map((c) => c.get(idx))).toEqual(expected[expected_idx++]);
-                            });
-                        });
-                        test(`calls bind function lazily`, () => {
-                            let bind = jest.fn();
-                            filtered.scan(() => { }, bind);
-                            if (expected.length) {
-                                expect(bind).toHaveBeenCalled();
-                            } else {
-                                expect(bind).not.toHaveBeenCalled();
-                            }
-                        });
-                    });
-                    describe(`scanReverse()`, () => {
-                        test(`iterates over expected values in reverse`, () => {
-                            let expected_idx = expected.length;
-                            filtered.scanReverse((idx, batch) => {
-                                const columns = batch.schema.fields.map((_, i) => batch.getChildAt(i)!);
-                                expect(columns.map((c) => c.get(idx))).toEqual(expected[--expected_idx]);
-                            });
-                        });
-                        test(`calls bind function lazily`, () => {
-                            let bind = jest.fn();
-                            filtered.scanReverse(() => { }, bind);
-                            if (expected.length) {
-                                expect(bind).toHaveBeenCalled();
-                            } else {
-                                expect(bind).not.toHaveBeenCalled();
-                            }
-                        });
-                    });
-                });
-            }
-            test(`countBy on dictionary returns the correct counts`, () => {
-                // Make sure countBy works both with and without the Col wrapper
-                // class
-                let expected: { [key: string]: number } = { 'a': 0, 'b': 0, 'c': 0 };
-                for (let row of values) {
-                    expected[row[DICT]] += 1;
-                }
 
-                expect(table.countBy(col('dictionary')).toJSON()).toEqual(expected);
-                expect(table.countBy('dictionary').toJSON()).toEqual(expected);
-            });
-            test(`countBy on dictionary with filter returns the correct counts`, () => {
-                let expected: { [key: string]: number } = { 'a': 0, 'b': 0, 'c': 0 };
-                for (let row of values) {
-                    if (row[I32] === 1) { expected[row[DICT]] += 1; }
-                }
-
-                expect(table.filter(col('i32').eq(1)).countBy('dictionary').toJSON()).toEqual(expected);
-            });
-            test(`countBy on non dictionary column throws error`, () => {
-                expect(() => { table.countBy('i32'); }).toThrow();
-                expect(() => { table.filter(col('dict').eq('a')).countBy('i32'); }).toThrow();
-            });
-            test(`countBy on non-existent column throws error`, () => {
-                expect(() => { table.countBy('FAKE' as any); }).toThrow();
-            });
             test(`table.select() basic tests`, () => {
                 let selected = table.select('f32', 'dictionary');
                 expect(selected.schema.fields).toHaveLength(2);
@@ -525,63 +345,9 @@ describe(`Table`, () => {
                     expect(row.dictionary).toEqual(expected_row[DICT]);
                 }
             });
-            // test(`table.toString()`, () => {
-            //     let selected = table.select('i32', 'dictionary');
-            //     let headers = [`"row_id"`, `"i32: Int32"`, `"dictionary: Dictionary<Int8, Utf8>"`];
-            //     let expected = [headers.join(' | '), ...values.map((row, idx) => {
-            //         return [`${idx}`, `${row[I32]}`, `"${row[DICT]}"`].map((str, col) => {
-            //             return leftPad(str, ' ', headers[col].length);
-            //         }).join(' | ');
-            //     })].join('\n') + '\n';
-            //     expect(selected.toString()).toEqual(expected);
-            // });
-            test(`table.filter(..).count() on always false predicates returns 0`, () => {
-                expect(table.filter(col('i32').ge(100)).count()).toEqual(0);
-                expect(table.filter(col('dictionary').eq('z')).count()).toEqual(0);
-            });
-            describe(`lit-lit comparison`, () => {
-                test(`always-false count() returns 0`, () => {
-                    expect(table.filter(lit('abc').eq('def')).count()).toEqual(0);
-                    expect(table.filter(lit(0).ge(1)).count()).toEqual(0);
-                });
-                test(`always-true count() returns length`, () => {
-                    expect(table.filter(lit('abc').eq('abc')).count()).toEqual(table.length);
-                    expect(table.filter(lit(-100).le(0)).count()).toEqual(table.length);
-                });
-            });
-            describe(`col-col comparison`, () => {
-                test(`always-false count() returns 0`, () => {
-                    expect(table.filter(col('dictionary').eq(col('i32'))).count()).toEqual(0);
-                });
-                test(`always-true count() returns length`, () => {
-                    expect(table.filter(col('dictionary').eq(col('dictionary'))).count()).toEqual(table.length);
-                });
-            });
         });
     }
 });
-
-describe(`Predicate`, () => {
-    const p1 = col('a').gt(100);
-    const p2 = col('a').lt(1000);
-    const p3 = col('b').eq('foo');
-    const p4 = col('c').eq('bar');
-    const expected = [p1, p2, p3, p4];
-    test(`and flattens children`, () => {
-        expect(and(p1, p2, p3, p4).children).toEqual(expected);
-        expect(and(p1.and(p2), new And(p3, p4)).children).toEqual(expected);
-        expect(and(p1.and(p2, p3, p4)).children).toEqual(expected);
-    });
-    test(`or flattens children`, () => {
-        expect(or(p1, p2, p3, p4).children).toEqual(expected);
-        expect(or(p1.or(p2), new Or(p3, p4)).children).toEqual(expected);
-        expect(or(p1.or(p2, p3, p4)).children).toEqual(expected);
-    });
-});
-
-// function leftPad(str: string, fill: string, n: number) {
-//     return (new Array(n + 1).join(fill) + str).slice(-1 * n);
-// }
 
 type TestDataSchema = { f32: Float32; i32: Int32; dictionary: Dictionary<Utf8, Int8> };
 
@@ -594,7 +360,7 @@ function getTestVectors(f32Values: number[], i32Values: number[], dictIndices: n
     return [Vector.new(f32Data), Vector.new(i32Data), DictionaryVector.from(values, new Int8(), dictIndices)];
 }
 
-export function getSingleRecordBatchTable() {
+function getSingleRecordBatchTable() {
     const vectors = getTestVectors(
         [-0.3, -0.2, -0.1, 0, 0.1, 0.2, 0.3],
         [-1, 1, -1, 1, -1, 1, -1],
